Added Perl::Use() / Perl::LoadModule() / Perl::Run()
svn path=/trunk/perl-bindings/; revision=10784

+++ src/Y2PerlComponent.cc
@@ -60,17 +60,20 @@ Y2PerlComponent::evaluate( const YCPValue & val )
 
     y2debug("YPerl::evaluate( %s, %s )", function.c_str(), argList->toString().c_str());
 
-    // Commands are ordered by how commonly they are called (optimization).
+    // Optimization: Commands are ordered by how frequently they are called.
     
     if ( function == "CallVoid" 	)	return YPerl::callVoid	( argList );
     if ( function == "CallList" 	)	return YPerl::callList	( argList );
+    if ( function == "CallString" 	)	return YPerl::callString( argList );
     if ( function == "CallBool" 	)	return YPerl::callBool	( argList );
     if ( function == "CallBoolean"	)	return YPerl::callBool	( argList );
-    if ( function == "CallString" 	)	return YPerl::callString( argList );
     if ( function == "CallInt" 		)	return YPerl::callInt	( argList );
     if ( function == "CallInteger" 	)	return YPerl::callInt	( argList );
     if ( function == "Eval" 		)	return YPerl::eval	( argList );
     if ( function == "Parse" 		)	return YPerl::parse	( argList );
+    if ( function == "Run" 		)	return YPerl::run	( argList );
+    if ( function == "LoadModule"	)	return YPerl::loadModule( argList );
+    if ( function == "Use"		)	return YPerl::loadModule( argList );
     if ( function == "Destroy" 		)	return YPerl::destroy();
 	
     return YCPError( string ( "Undefined Perl::" ) + function );
